Fix type issues
